fix: (내) 풀이 상세 코드창 분리
