+++ frontend/Prog-rangers-main/src/components/MySolutionDetail/MySolView.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { css } from '@mui/material';
 import axios from 'axios';
 import { useParams } from 'react-router-dom';
-import { CodeWindow2 } from '../Profile';
+import { CodeWindow3 } from '../Profile';
 import {
   LineContentMock,
   contentLayout,
@@ -75,7 +75,7 @@ export const MySolView = () => {
             `}
           ></div>
 
-          <CodeWindow2 />
+          <CodeWindow3 />
         </div>
       </div>
     </>
