+++ frontend/Prog-rangers-main/src/components/Profile/index.js
@@ -3,3 +3,4 @@ export { CodeWindow } from './CodeWindow';
 export { CodeWindow2 } from './CodeWindow2';
 export { CodeWindow3 } from './CodeWindow3';
 export { CodeWindow4 } from './CodeWindow4';
+export { CodeWindow5 } from './CodeWindow5';
